Refactoring knockoff part + improve docstring
