+++ hidimstat/__init__.py
@@ -2,11 +2,14 @@ from .clustered_inference import clustered_inference, hd_inference
 from .desparsified_lasso import desparsified_lasso_confint
 from .ensemble_clustered_inference import ensemble_clustered_inference
 from .gaonkar import gaonkar
+from .gaussian_knockoff import gaussian_knockoff_generation
+from .knockoffs import knockoff_aggregation, model_x_knockoff
 from .multi_sample_split import aggregate_medians, aggregate_quantiles
 from .noise_std import reid
 from .permutation_test import permutation_test, permutation_test_cv
 from .scenario import design_matrix_toeplitz_cov, scenario
 from .standardized_svr import standardized_svr
+from .stat_coef_diff import stat_coef_diff
 from .version import __version__
 
 __all__ = [
@@ -17,10 +20,14 @@ __all__ = [
     'desparsified_lasso_confint',
     'ensemble_clustered_inference',
     'gaonkar',
+    'gaussian_knockoff_generation',
     'hd_inference',
+    'knockoff_aggregation',
+    'model_x_knockoff',
     'permutation_test',
     'permutation_test_cv',
     'reid',
+    'stat_coef_diff',
     'scenario',
     'standardized_svr',
     '__version__',
