+++ hidimstat/stat_coef_diff.py
