+++ hidimstat/test/test_knockoff_aggregation.py
@@ -1,11 +1,11 @@
 import numpy as np
+from hidimstat.data_simulation import simu_data
 from hidimstat.knockoffs import knockoff_aggregation, model_x_knockoff
-from hidimstat.knockoffs.data_simulation import simu_data
-from hidimstat.knockoffs.utils import cal_fdp_power
+from hidimstat.utils import cal_fdp_power
 
 n = 600
 p = 1000
-n_bootstraps = 25
+n_bootstraps = 15
 fdr = 0.1
 SEED = 0
