+++ hidimstat/test/test_stat_coef_diff.py
